feat: combine test cases
