+++ data_structures/test/singly_linked_list.test.ts
@@ -1,0 +1,4 @@
+import { SinglyLinkedList } from "../singly_linked_list";
+import { testLinkedList } from "./linked_list";
+
+describe("Singly linked list", () => testLinkedList(SinglyLinkedList));
